+++ src/Firebase/Auth/SignInWithIdpCredentials.php
@@ -15,6 +15,9 @@ final class SignInWithIdpCredentials implements SignIn
     /** @var string */
     private $provider;
 
+    /** @var string|null */
+    private $oauthTokenSecret;
+
     /** @var string */
     private $requestUri = 'http://localhost';
 
@@ -27,6 +30,14 @@ final class SignInWithIdpCredentials implements SignIn
         $instance = new self();
         $instance->provider = $provider;
         $instance->accessToken = $accessToken;
+
+        return $instance;
+    }
+
+    public static function withAccessTokenAndOauthTokenSecret(string $provider, string $accessToken, string $oauthTokenSecret): self
+    {
+        $instance = self::withAccessToken($provider, $accessToken);
+        $instance->oauthTokenSecret = $oauthTokenSecret;
 
         return $instance;
     }
@@ -51,6 +62,11 @@ final class SignInWithIdpCredentials implements SignIn
     public function provider(): string
     {
         return $this->provider;
+    }
+
+    public function oauthTokenSecret(): ?string
+    {
+        return $this->oauthTokenSecret;
     }
 
     public function accessToken(): ?string